Update UI, eraseInput method, sqlinjection method connecting the program to db
