+++ src/Employees.form
@@ -1,5 +1,5 @@
 <?xml version="1.0" encoding="UTF-8"?>
-<form xmlns="http://www.intellij.com/uidesigner/form/" version="1" bind-to-class="employees">
+<form xmlns="http://www.intellij.com/uidesigner/form/" version="1" bind-to-class="Employees">
   <grid id="27dc6" binding="mainPanel" layout-manager="GridLayoutManager" row-count="9" column-count="6" same-size-horizontally="false" same-size-vertically="false" hgap="-1" vgap="-1">
     <margin top="0" left="0" bottom="0" right="0"/>
     <constraints>
@@ -11,17 +11,7 @@
     </properties>
     <border type="none"/>
     <children>
-      <component id="797b8" class="javax.swing.JLabel" binding="salary">
-        <constraints>
-          <grid row="5" column="1" row-span="1" col-span="1" vsize-policy="0" hsize-policy="0" anchor="8" fill="0" indent="0" use-parent-layout="false"/>
-        </constraints>
-        <properties>
-          <enabled value="true"/>
-          <foreground color="-16777216"/>
-          <text value="Salary:"/>
-        </properties>
-      </component>
-      <component id="9c48b" class="javax.swing.JLabel" binding="contactNumber">
+      <component id="9c48b" class="javax.swing.JLabel">
         <constraints>
           <grid row="4" column="1" row-span="1" col-span="1" vsize-policy="0" hsize-policy="0" anchor="8" fill="0" indent="0" use-parent-layout="false"/>
         </constraints>
@@ -30,7 +20,7 @@
           <text value="Contact Number:"/>
         </properties>
       </component>
-      <component id="7b0d" class="javax.swing.JLabel" binding="lastName">
+      <component id="7b0d" class="javax.swing.JLabel">
         <constraints>
           <grid row="3" column="1" row-span="1" col-span="1" vsize-policy="0" hsize-policy="0" anchor="8" fill="0" indent="0" use-parent-layout="false"/>
         </constraints>
@@ -39,7 +29,7 @@
           <text value="Last Name:"/>
         </properties>
       </component>
-      <component id="2e3f2" class="javax.swing.JLabel" binding="firstName">
+      <component id="2e3f2" class="javax.swing.JLabel">
         <constraints>
           <grid row="2" column="1" row-span="1" col-span="1" vsize-policy="0" hsize-policy="0" anchor="8" fill="0" indent="0" use-parent-layout="false"/>
         </constraints>
@@ -48,7 +38,7 @@
           <text value="First Name:"/>
         </properties>
       </component>
-      <component id="5a10a" class="javax.swing.JLabel" binding="employeeId">
+      <component id="5a10a" class="javax.swing.JLabel">
         <constraints>
           <grid row="1" column="1" row-span="1" col-span="1" vsize-policy="0" hsize-policy="0" anchor="8" fill="0" indent="0" use-parent-layout="false"/>
         </constraints>
@@ -206,6 +196,16 @@
         <border type="none"/>
         <children/>
       </grid>
+      <component id="797b8" class="javax.swing.JLabel">
+        <constraints>
+          <grid row="5" column="1" row-span="1" col-span="1" vsize-policy="0" hsize-policy="0" anchor="8" fill="0" indent="0" use-parent-layout="false"/>
+        </constraints>
+        <properties>
+          <enabled value="true"/>
+          <foreground color="-16777216"/>
+          <text value="Salary:"/>
+        </properties>
+      </component>
     </children>
   </grid>
 </form>
